+++ proxy/public/index.html
@@ -1,0 +1,39 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+  <meta charset="UTF-8">
+  <style type="text/css">
+    @font-face {
+      font-family: 'DIN Pro';
+      font-style: normal;
+      src: url(https://d2ue93q3u507c2.cloudfront.net/assets/generated_assets/e0be9bd5e372b2c379893b776e7b2b96.otf) format("opentype");
+    }
+    body {
+      font-family: 'DIN Pro';
+      background: #1B1B1C;
+    }
+  </style>
+  <link href="https://d2ue93q3u507c2.cloudfront.net/assets/favicon.ico" rel="shortcut icon">
+  <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css">
+  <link rel="stylesheet" href="overview.css">
+  <script src="https://code.jquery.com/jquery-3.3.1.slim.min.js" integrity="sha256-3edrmyuQ0w65f8gfBsqowzjJe2iM6n0nKciPUp8y+7E="
+  crossorigin="anonymous"></script>
+  <title>Robinhood</title>
+</head>
+
+<body>
+  <div id='overview'>
+    <div id="stockPriceChart"></div>
+    <div id="buyServiceContainer">
+      <div id="buyService"></div>
+    </div>
+    <div id="priceVolumeChart"></div>
+    <div id="app"></div>
+  </div>
+  <script src="https://s3.amazonaws.com/stockpricechart/bundle.js"></script>
+  <script src="https://s3-us-west-1.amazonaws.com/tson-rh-buy-service/bundle.js"></script>
+  <script src="https://s3.us-east-2.amazonaws.com/price-volume-chart/bundle.js"></script>
+  <script src="https://s3-us-west-1.amazonaws.com/robinhoodfecben/bundle.js"></script>
+</body>
+</html>
