+++ proxy/server.js
@@ -1,0 +1,52 @@
+const express = require('express');
+const path = require('path');
+const httpProxy = require('http-proxy');
+
+const app = express();
+const apiProxy = httpProxy.createProxyServer();
+
+const stockPriceChart = 'http://ec2-35-172-183-27.compute-1.amazonaws.com/',
+      priceVolumeChart = 'http://ec2-34-224-212-238.compute-1.amazonaws.com/',
+      buyService = 'http://ec2-34-200-253-32.compute-1.amazonaws.com/',
+      peopleAlsoBought = 'http://ec2-18-224-182-229.us-east-2.compute.amazonaws.com/';
+
+
+app.use(express.static(__dirname + '/public'));
+const port = 3000;
+
+let stockPriceChartId = 0;
+app.all('/api/symbol/:stockPriceChartId/day', (req, res) => {
+  stockPriceChartId = req.params.stockPriceChartId;
+  apiProxy.web(req, res, {target: stockPriceChart});
+});
+
+app.all(`/api/symbol/:${stockPriceChartId + 1}/week`, (req, res) => {
+  apiProxy.web(req, res, { target: stockPriceChart });
+});
+
+app.all(`/api/symbol/:${stockPriceChartId + 2}/week`, (req, res) => {
+  apiProxy.web(req, res, { target: stockPriceChart });
+});
+
+app.all(`/api/symbol/:${stockPriceChartId + 3}/week`, (req, res) => {
+  apiProxy.web(req, res, { target: stockPriceChart });
+});
+
+app.all(`/api/symbol/:${stockPriceChartId + 4}/week`, (req, res) => {
+  apiProxy.web(req, res, { target: stockPriceChart });
+});
+
+app.all(`/api/volumes/symbols/:id`, (req, res) => {
+  apiProxy.web(req, res, { target: priceVolumeChart });
+});
+
+app.all('/api/buytest', (req, res) => {
+  apiProxy.web(req, res, { target: buyService });
+});
+
+app.all('/api/alsoBought/1', (req, res) => {
+  // console.log(req.);
+  apiProxy.web(req, res, { target: peopleAlsoBought });
+});
+
+app.listen(port, () => console.log('Server listening on port ' + port));
