--- public/index.html
@@ -1,34 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <meta charset="utf-8" />
-    <title>Page Title</title>
-    <script crossorigin src="https://unpkg.com/react@16/umd/react.development.js"></script>
-    <script crossorigin src="https://unpkg.com/react-dom@16/umd/react-dom.development.js"></script>
-    <style type="text/css">
-        @font-face {
-          font-family: 'DIN Pro';
-          font-style: normal;
-          src: url(https://d2ue93q3u507c2.cloudfront.net/assets/generated_assets/e0be9bd5e372b2c379893b776e7b2b96.otf) format("opentype");
-        }
-        body {
-          font-family: 'DIN Pro';
-        }
-      </style>
-</head>
-<body>
-    <div id="stockPriceChart"></div>
-    <div id="buyService"></div>
-    <div id="priceVolumeChart"></div>
-    <div id="app"></div>
-
-    <script
-  src="https://code.jquery.com/jquery-3.3.1.js"
-  integrity="sha256-2Kok7MbOyxpgUVvAk/HJ2jigOSYS2auK4Pfzbm7uH60="
-  crossorigin="anonymous"></script>
-    <script src="https://s3-us-west-1.amazonaws.com/robinhoodfecben/bundle.js"></script>
-    <script src="https://s3.us-east-2.amazonaws.com/price-volume-chart/bundle.js"></script>
-    <script src="https://s3.amazonaws.com/stockpricechart/bundle.js"></script>
-    <script src="https://s3-us-west-1.amazonaws.com/tson-rh-buy-service/bundle.js"></script>
-</body>
-</html>
